+++ docs-src/pages/changelog.svelte
@@ -1,7 +1,7 @@
 <h1>Changelog</h1>
 <h2>v9.1.0 <em>(2023-09-?)</em></h2>
 <ul>
-<li>New components: <code>InputRating</code>, <code>Tag</code>, <code>InputTag</code>.</li>
+<li>New components: <code>InputRating</code>, <code>Tag</code>, <code>InputTag</code>, <code>InputTime</code>.</li>
 <li>Add <code>hideTip</code> and more, to <code>Popover</code>.</li>
 <li>Small bugfixes and improvements.</li>
 </ul>