+++ src/input/index.js
@@ -9,6 +9,7 @@ export * from './input-password';
 export * from './input-rating';
 export * from './input-search';
 export * from './input-tag';
+export * from './input-time';
 export * from './input-text';
 export * from './label';
 export * from './radio';